requirements.txt
datascience
numpydoc
sphinx
pytest-cov